+++ radix-transactions/src/model/preparation/traits.rs
@@ -7,7 +7,7 @@ pub trait TransactionPayload:
     + ManifestDecode
     + ManifestCategorize
     + for<'a> ManifestSborEnumVariantFor<
-        VersionedTransactionPayload,
+        AnyTransaction,
         OwnedVariant: ManifestDecode,
         BorrowedVariant<'a>: ManifestEncode,
     >
